Modularized the project.
Current Modules:-
 - modules.h
 - problem.h
 - heuristics.h
