+++ heuristics.h
@@ -1,0 +1,6 @@
+
+int get_row(int x,int s);
+int get_column(int x,int s);
+int manhattan(int *board,int s);
+int euc_sq(int *board,int s);
+int max(int *board, int s);
